+++ Changes.md
@@ -14,8 +14,8 @@ Improvements
 
 - Spreadsheet :
   - Enabled cell selection, editing is now achieved with a double-click.
-  - Added menu items to copy/paste the enabled state and value(s) of selected cells.
-  - Added menu items to delete the selected rows.
+  - Added menu items to cells & defaults sections to copy/paste the enabled state and value(s) of selected cells.
+  - Added menu items to the row names section to copy/paste or delete the selected rows.
   - Added menu items to edit the values for all selected cells of the same type simultaneously.
 
 Fixes